requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0
django-bootstrap5==23.3
django-crispy-forms==2.1
mysql-connector-python==8.2.0
mysqlclient==2.2.1
Pillow==10.1.0
protobuf==4.21.12
sqlparse==0.4.4
typing_extensions==4.9.0
six
django-recaptcha
python-decouple